bergen/requirements.txt
Django>=2.0.0
channels>=2.0.2
channels_redis~=2.0
django-registration-redux
django-autocomplete-light
django-taggit
django-extensions

jupyter

nibabel
SimpleITK
django-cors-headers
django-crispy-forms
django-filter
django-pandas
djangorestframework
django-oauth-toolkit
matplotlib

docutils
et-xmlfile
future
h5py
jmespath
opencv-python
pandas
Pillow
python-dateutil
pytz
rest-pandas
s3transfer
six
wincertstore
boto3
botocore
certifi

beautifulsoup4
lxml

javabridge
python-bioformats

scipy
django-mptt
easy_thumbnails
django-polymorphic
psycopg2
